Port the shmem_pipe transport to work with Phoenix.
This mainly consists of adding nonblocking mode functionality, as well as conforming to expected unix-y behaviours,
like returning smaller buffers than were asked for if the initial request can't be fulfilled but some space is available.

To come: close, deal with lent buffers (just a generic wrapper; lenders are no use to us), and debug.